requirements.txt
argparse
firecrawl-py
langchain_text_splitters
pdfplumber
pymilvus
pymilvus.model
openai
numpy
tqdm
termcolor
fastapi
uvicorn
pydantic-settings
jieba>=0.42.1
pymysql>=1.1.0
PyYAML>=6.0.1
requests>=2.31.0
urllib3>=2.0.7
langchain-core>=0.1.0
oracledb>=1.4.2
aiomysql>=0.2.0
alibabacloud_green20220302==2.20.0
langid>=1.1.6
inflect>=7.5.0
xlwt>=1.3.0
aliyun-mns-sdk>=1.2.0